consumers/hotspot_consumer/requirements.txt
kafka-python==2.0.2
redis==4.5.5
python-dotenv==1.0.0
loguru==0.7.2
